Feature/login

+++ src/main/java/com/project/manageus/repository/UserRepository.java
@@ -3,7 +3,5 @@ package com.project.manageus.repository;
 import com.project.manageus.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-
-
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 }
